Move PaymentAuth3ds2ChallengeStatusReceiver to its own class
This helps make `StripePaymentController` more manageable.
