+++ stripe/src/test/java/com/stripe/android/StripePaymentControllerTest.kt
@@ -15,7 +15,6 @@ import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
-import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
@@ -29,7 +28,6 @@ import com.stripe.android.model.PaymentIntentFixtures
 import com.stripe.android.model.SetupIntentFixtures
 import com.stripe.android.model.Source
 import com.stripe.android.model.SourceFixtures
-import com.stripe.android.model.Stripe3ds2AuthResult
 import com.stripe.android.model.Stripe3ds2AuthResultFixtures
 import com.stripe.android.model.Stripe3ds2Fingerprint
 import com.stripe.android.model.Stripe3ds2FingerprintTest
@@ -41,18 +39,10 @@ import com.stripe.android.networking.AnalyticsDataFactory
 import com.stripe.android.networking.AnalyticsRequest
 import com.stripe.android.networking.AnalyticsRequestExecutor
 import com.stripe.android.networking.ApiRequest
-import com.stripe.android.networking.RetryDelaySupplier
 import com.stripe.android.payments.PaymentFlowResult
-import com.stripe.android.payments.Stripe3ds2CompletionContract
-import com.stripe.android.payments.Stripe3ds2CompletionStarter
 import com.stripe.android.stripe3ds2.service.StripeThreeDs2Service
-import com.stripe.android.stripe3ds2.transaction.ChallengeFlowOutcome
 import com.stripe.android.stripe3ds2.transaction.ChallengeParameters
-import com.stripe.android.stripe3ds2.transaction.CompletionEvent
-import com.stripe.android.stripe3ds2.transaction.ErrorMessage
 import com.stripe.android.stripe3ds2.transaction.MessageVersionRegistry
-import com.stripe.android.stripe3ds2.transaction.ProtocolErrorEvent
-import com.stripe.android.stripe3ds2.transaction.RuntimeErrorEvent
 import com.stripe.android.stripe3ds2.transaction.SdkTransactionId
 import com.stripe.android.stripe3ds2.transaction.Stripe3ds2ActivityStarterHost
 import com.stripe.android.stripe3ds2.transaction.Transaction
@@ -115,11 +105,6 @@ internal class StripePaymentControllerTest {
 
     private val paymentAuthWebViewContract = PaymentAuthWebViewContract()
     private val controller = createController()
-
-    private val completionStarter = Stripe3ds2CompletionStarter.Legacy(
-        host,
-        50000
-    )
 
     @BeforeTest
     fun setup() {
@@ -341,227 +326,6 @@ internal class StripePaymentControllerTest {
     }
 
     @Test
-    fun test3ds2Receiver_whenCompleted_shouldFireAnalyticsRequest() {
-        val completionEvent = CompletionEvent(
-            sdkTransactionId = sdkTransactionId,
-            transactionStatus = "C"
-        )
-
-        whenever(transaction.initialChallengeUiType).thenReturn("04")
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.completed(
-            completionEvent,
-            "01",
-            ChallengeFlowOutcome.CompleteSuccessful
-        )
-
-        verify(analyticsRequestExecutor, times(2))
-            .executeAsync(analyticsRequestArgumentCaptor.capture())
-        val analyticsRequests = analyticsRequestArgumentCaptor.allValues
-
-        assertThat(requireNotNull(analyticsRequests[0].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengeCompleted.toString())
-
-        val analyticsParamsSecond = requireNotNull(analyticsRequests[1].params)
-        assertThat(analyticsParamsSecond[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengePresented.toString())
-        assertThat(analyticsParamsSecond[AnalyticsDataFactory.FIELD_3DS2_UI_TYPE])
-            .isEqualTo("oob")
-    }
-
-    @Test
-    fun test3ds2Receiver_whenTimedout_shouldFireAnalyticsRequest() {
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.timedout("01")
-        verify(analyticsRequestExecutor, times(2))
-            .executeAsync(analyticsRequestArgumentCaptor.capture())
-        val analyticsRequests = analyticsRequestArgumentCaptor.allValues
-
-        assertThat(requireNotNull(analyticsRequests[0].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengeTimedOut.toString())
-
-        assertThat(requireNotNull(analyticsRequests[1].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengePresented.toString())
-    }
-
-    @Test
-    fun test3ds2Receiver_whenCanceled_shouldFireAnalyticsRequest() {
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.cancelled("01")
-
-        verify(analyticsRequestExecutor, times(2))
-            .executeAsync(analyticsRequestArgumentCaptor.capture())
-        val analyticsRequests = analyticsRequestArgumentCaptor.allValues
-
-        assertThat(requireNotNull(analyticsRequests[0].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengeCanceled.toString())
-
-        assertThat(requireNotNull(analyticsRequests[1].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengePresented.toString())
-    }
-
-    @Test
-    fun test3ds2Receiver_whenRuntimeErrorError_shouldFireAnalyticsRequest() {
-        val runtimeErrorEvent = RuntimeErrorEvent(
-            errorCode = "404",
-            errorMessage = "Resource not found"
-        )
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.runtimeError(runtimeErrorEvent)
-
-        verify(analyticsRequestExecutor, times(2))
-            .executeAsync(analyticsRequestArgumentCaptor.capture())
-        val analyticsRequests = analyticsRequestArgumentCaptor.allValues
-
-        val analyticsParamsFirst = requireNotNull(analyticsRequests[0].params)
-        assertThat(analyticsParamsFirst[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengeErrored.toString())
-
-        assertThat(analyticsParamsFirst[AnalyticsDataFactory.FIELD_ERROR_DATA])
-            .isEqualTo(
-                mapOf(
-                    "type" to "runtime_error_event",
-                    "error_code" to "404",
-                    "error_message" to "Resource not found"
-                )
-            )
-
-        assertThat(requireNotNull(analyticsRequests[1].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengePresented.toString())
-    }
-
-    @Test
-    fun test3ds2Receiver_whenProtocolError_shouldFireAnalyticsRequest() {
-        val protocolErrorEvent = ProtocolErrorEvent(
-            sdkTransactionId = sdkTransactionId,
-            errorMessage = ErrorMessage(
-                errorCode = "201",
-                errorDescription = "Required element missing",
-                errorDetails = "eci",
-                transactionId = "047f76a6-d1d4-48a2-aa65-786abb6f7f46"
-            )
-        )
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.protocolError(protocolErrorEvent)
-
-        verify(analyticsRequestExecutor, times(2))
-            .executeAsync(analyticsRequestArgumentCaptor.capture())
-        val analyticsRequests = analyticsRequestArgumentCaptor.allValues
-
-        val analyticsParamsFirst = requireNotNull(analyticsRequests[0].params)
-        assertThat(analyticsParamsFirst[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengeErrored.toString())
-
-        assertThat(analyticsParamsFirst[AnalyticsDataFactory.FIELD_ERROR_DATA])
-            .isEqualTo(
-                mapOf(
-                    "type" to "protocol_error_event",
-                    "error_code" to "201",
-                    "sdk_trans_id" to sdkTransactionId.value,
-                    "error_description" to "Required element missing",
-                    "error_details" to "eci",
-                    "trans_id" to "047f76a6-d1d4-48a2-aa65-786abb6f7f46"
-                )
-            )
-
-        assertThat(requireNotNull(analyticsRequests[1].params)[AnalyticsDataFactory.FIELD_EVENT])
-            .isEqualTo(AnalyticsEvent.Auth3ds2ChallengePresented.toString())
-    }
-
-    @Test
-    fun test3ds2Completion_whenCanceled_shouldCallStarterWithCancelStatus() {
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-            workContext = testDispatcher
-        )
-        receiver.cancelled("01")
-
-        verify(activity).startActivityForResult(
-            intentArgumentCaptor.capture(),
-            eq(50000)
-        )
-
-        val paymentFlowResult = Stripe3ds2CompletionContract().parsePaymentFlowResult(
-            intentArgumentCaptor.firstValue
-        )
-
-        assertThat(paymentFlowResult)
-            .isEqualTo(
-                PaymentFlowResult.Unvalidated(
-                    clientSecret = PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2.clientSecret.orEmpty(),
-                    stripeAccountId = ACCOUNT_ID,
-                    flowOutcome = StripeIntentResult.Outcome.CANCELED
-                )
-            )
-    }
-
-    @Test
     fun handlePaymentResult_withAuthException_shouldCallCallbackOnError() {
         val exception = APIException(RuntimeException())
         val intent = Intent().putExtras(
@@ -770,8 +534,8 @@ internal class StripePaymentControllerTest {
             )
 
             verify(paymentRelayStarter).start(
-                argWhere<PaymentRelayStarter.Args.ErrorArgs> { args ->
-                    args.exception.message == "Error while attempting to start 3DS2 challenge flow."
+                argWhere<PaymentRelayStarter.Args.ErrorArgs> {
+                    it.exception.message == "Error while attempting to start 3DS2 challenge flow."
                 }
             )
         }
@@ -962,141 +726,6 @@ internal class StripePaymentControllerTest {
                     shouldCancelSource = true
                 )
             )
-    }
-
-    @Test
-    fun `complete3ds2Auth() should retry until max retries are attempted due to a 4xx response`() {
-        var complete3ds2AuthInvocations = 0
-        val stripeRepository = object : AbsFakeStripeRepository() {
-            override suspend fun complete3ds2Auth(
-                sourceId: String,
-                requestOptions: ApiRequest.Options
-            ): Stripe3ds2AuthResult {
-                complete3ds2AuthInvocations++
-
-                throw InvalidRequestException(
-                    statusCode = 400
-                )
-            }
-        }
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-
-            // set to 0 so there is effectively no delay
-            retryDelaySupplier = RetryDelaySupplier(incrementSeconds = 0L),
-
-            workContext = testDispatcher
-        )
-
-        receiver.timedout("1")
-
-        assertThat(complete3ds2AuthInvocations)
-            .isEqualTo(4)
-    }
-
-    @Test
-    fun `complete3ds2Auth() should succeed after a single retry failure due to a 4xx response`() {
-        var complete3ds2AuthInvocations = 0
-
-        val stripeRepository = object : AbsFakeStripeRepository() {
-            override suspend fun complete3ds2Auth(
-                sourceId: String,
-                requestOptions: ApiRequest.Options
-            ): Stripe3ds2AuthResult {
-                complete3ds2AuthInvocations++
-                if (complete3ds2AuthInvocations <= 2) {
-                    throw InvalidRequestException(
-                        statusCode = 400
-                    )
-                } else {
-                    return Stripe3ds2AuthResultFixtures.CHALLENGE_COMPLETION
-                }
-            }
-        }
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-
-            // set to 0 so there is effectively no delay
-            retryDelaySupplier = RetryDelaySupplier(incrementSeconds = 0L),
-
-            workContext = testDispatcher
-        )
-
-        receiver.timedout("1")
-
-        assertThat(complete3ds2AuthInvocations)
-            .isEqualTo(3)
-
-        verify(activity).startActivityForResult(
-            argWhere { intent ->
-                val result = requireNotNull(
-                    Stripe3ds2CompletionContract().parsePaymentFlowResult(intent)
-                )
-                result == PaymentFlowResult.Unvalidated(
-                    clientSecret = "pi_1ExkUeAWhjPjYwPiXph9ouXa_secret_nGTdfGlzL9Uop59wN55LraiC7",
-                    flowOutcome = StripeIntentResult.Outcome.TIMEDOUT,
-                    stripeAccountId = ACCOUNT_ID
-                )
-            },
-            eq(StripePaymentController.PAYMENT_REQUEST_CODE)
-        )
-    }
-
-    @Test
-    fun `complete3ds2Auth() should not retry after a 5xx response`() {
-        var complete3ds2AuthInvocations = 0
-        val stripeRepository = object : AbsFakeStripeRepository() {
-            override suspend fun complete3ds2Auth(
-                sourceId: String,
-                requestOptions: ApiRequest.Options
-            ): Stripe3ds2AuthResult {
-                complete3ds2AuthInvocations++
-
-                throw APIException(
-                    statusCode = 500
-                )
-            }
-        }
-
-        val receiver = StripePaymentController.PaymentAuth3ds2ChallengeStatusReceiver(
-            completionStarter,
-            stripeRepository,
-            PaymentIntentFixtures.PI_REQUIRES_MASTERCARD_3DS2,
-            "src_123",
-            REQUEST_OPTIONS,
-            analyticsRequestExecutor,
-            analyticsDataFactory,
-            transaction,
-            AnalyticsRequest.Factory(),
-
-            // set to 0 so there is effectively no delay
-            retryDelaySupplier = RetryDelaySupplier(incrementSeconds = 0L),
-
-            workContext = testDispatcher
-        )
-
-        receiver.timedout("1")
-
-        assertThat(complete3ds2AuthInvocations)
-            .isEqualTo(1)
     }
 
     private fun createController(
